backend/requirements.txt
# Backend API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# PDF Generation
reportlab>=4.0.0

# File handling
aiofiles>=23.0.0

# CORS
python-multipart>=0.0.6